Move files to kit
Move some files to kit that should have been there in the first
place.
