+++ kit/src/main/java/com/oracle/javafx/scenebuilder/kit/library/IExternalLibraryItemProvider.java
@@ -1,0 +1,9 @@
+package com.oracle.javafx.scenebuilder.kit.library;
+
+import java.util.Collection;
+
+public interface IExternalLibraryItemProvider {
+
+	public Collection<Class<?>> getItems();
+	
+}
