+++ kit/src/main/resources/com/oracle/javafx/scenebuilder/kit/editor/panel/css/MessageBar.css
@@ -1,0 +1,67 @@
+/*
+ * Copyright (c) 2014, Oracle and/or its affiliates.
+ * All rights reserved. Use is subject to license terms.
+ *
+ * This file is available and licensed under the following license:
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ *
+ *  - Redistributions of source code must retain the above copyright
+ *    notice, this list of conditions and the following disclaimer.
+ *  - Redistributions in binary form must reproduce the above copyright
+ *    notice, this list of conditions and the following disclaimer in
+ *    the documentation and/or other materials provided with the distribution.
+ *  - Neither the name of Oracle Corporation nor the names of its
+ *    contributors may be used to endorse or promote products derived
+ *    from this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+ * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+ * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+ * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+ * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+ * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+ * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+.message-bar {
+    -fx-padding: 0;
+    -fx-spacing: 0;
+}
+
+/*.message-bar .button, .message-bar .label {*/
+.message-bar .label {
+    -fx-font-size: 0.875em;
+    -fx-background-color: transparent;
+    -fx-background-radius: 0;
+    -fx-label-padding: 1 5 0 5;
+    -fx-background-insets: 0;
+}
+
+.warning-button {
+    -fx-font-size: 0.667em;
+    -fx-shape: "M1.347,18C1,18,0.678,17.82,0.496,17.526c-0.182-0.295-0.199-0.663-0.044-0.973l8-16C8.622,0.214,8.968,0,9.347,0
+    s0.725,0.214,0.895,0.553l8,16c0.155,0.31,0.138,0.678-0.044,0.973C18.015,17.82,17.693,18,17.347,18H1.347z";
+    -fx-background-color: black, yellow;
+    -fx-background-insets: 0, 1 1 0.7 1;
+    -fx-padding: 6 2 2 2;
+    -fx-min-width: 20;
+    -fx-max-width: 20;
+    -fx-min-height: 19;
+    -fx-max-height: 19;
+}
+
+.message-bar .message-warning {
+    -fx-background-color: 
+        linear-gradient( to bottom, derive(-sb-warning-panel, 20%) 0%, -sb-warning-panel 100% );
+}
+
+.message-bar .message-info {
+    -fx-background-color: linear-gradient( to bottom, derive(-sb-message-panel, 20%), -sb-message-panel );
+}
