+++ kit/src/main/resources/com/oracle/javafx/scenebuilder/kit/i18n/SceneBuilderKit.properties
@@ -1,4 +1,4 @@
-# Copyright (c) 2016, 2019, Gluon and/or its affiliates.
+# Copyright (c) 2016, 2017, Gluon and/or its affiliates.
 # Copyright (c) 2012, 2014, Oracle and/or its affiliates.
 # All rights reserved. Use is subject to license terms.
 #
@@ -66,10 +66,15 @@ label.qualifier.horizontal = (horizontal)
 label.qualifier.vertical = (vertical)
 
 # -----------------------------------------------------------------------------
+# Message bar
+# -----------------------------------------------------------------------------
+message.bar.file.dirty = File has unsaved changes
+message.bar.details = Click to access the list of warning messages
+
+
+# -----------------------------------------------------------------------------
 # Log Messages
 # -----------------------------------------------------------------------------
-log.info.explore.end = End exploring {0}
-log.info.explore.folder = Start exploring FOLDER {0}
 log.info.explore.jar = Start exploring JAR {0}
 log.warning.inline.edit.internationalized.strings = Can''t inline edit internationalized strings
 log.warning.color.creation.error.hexadecimal = Can''t create color for hexadecimal value ''{0}''
@@ -89,10 +94,12 @@ library.panel.menu.view.sections = View as Sections
 # -----------------------------------------------------------------------------
 # Hierarchy Panel
 # -----------------------------------------------------------------------------
+hierarchy = Hierarchy
 hierarchy.placeholder.insert = insert 
 hierarchy.placeholder.insert.graphic = insert graphic
 hierarchy.column.header.classname = Class Name
 hierarchy.column.header.info = Info
+hierarchy.displays = Hierarchy displays
 hierarchy.displayoption.info = Info
 hierarchy.displayoption.fxid = fx:id
 hierarchy.displayoption.nodeid = Node Id
@@ -102,17 +109,9 @@ hierarchy.unresolved.resource = Unresolved resource
 hierarchy.unsupported.expression = Unsupported by Scene Builder
 
 # -----------------------------------------------------------------------------
-# Info panel
+# Controller
 # -----------------------------------------------------------------------------
-info.label.assigned = Assigned fx:id
-info.label.component = Component
-info.label.use.fx.root = Use fx:root construct
-# Used after number 1
-info.label.item = item
-# Used after any number greater than 1
-info.label.items = items
-# The parameter is the name of a Java class
-info.tooltip.controller = Your Controller class must extend {0}
+controller = Controller
 
 # -----------------------------------------------------------------------------
 # Content Panel
@@ -127,6 +126,9 @@ content.log.cannot.display.selected.n = {0} of {1} selected objects cannot be di
 # -----------------------------------------------------------------------------
 # Inspector Panel
 # -----------------------------------------------------------------------------
+inspector = Inspector
+inspector.by.property.name = View by property name
+inspector.by.property.type = View by property type
 inspector.cursor.chooseimage = Choose Image...
 inspector.cursor.inherited = Inherited (Default)
 inspector.cursor.inheritedparent = Inherited from Parent (Default)
@@ -172,11 +174,14 @@ inspector.rectangle2D.not.defined = not defined
 inspector.select.css.filter = Style Sheet files
 inspector.select.css.title = Add Style Sheet
 inspector.select.image = Image files
+inspector.show.all = Show All
+inspector.show.edited = Show Edited
 inspector.style.parsingerror = CSS Error parsing in-line style:  ''{0}''
 inspector.style.valuetypeerror = CSS value type error:  ''{0}''
 inspector.stylesheet.alreadyexist = Stylesheet {0} has already been added
 inspector.stylesheet.cannotopen = Cannot open {0}
 inspector.stylesheet.cannotreveal = Cannot reveal {0}
+inspector.view.sections = View Sections
 inspector.css.tooltip = The value for this property is overridden by CSS
 inspector.css.showcss = Show CSS
 inspector.css.overridden = The value for the property {0} is overridden by CSS
@@ -240,6 +245,19 @@ include.file.failed = Failed to include ''{0}''
 include.file.failed.target = Failed to include ''{0}'' under ''{1}''
 
 # -----------------------------------------------------------------------------
+# Info panel
+# -----------------------------------------------------------------------------
+info.label.assigned = Assigned fx:id
+info.label.component = Component
+info.label.use.fx.root = Use fx:root construct
+# Used after number 1
+info.label.item = item
+# Used after any number greater than 1
+info.label.items = items
+# The parameter is the name of a Java class
+info.tooltip.controller = Your Controller class must extend {0}
+
+# -----------------------------------------------------------------------------
 # Library Dialog
 # -----------------------------------------------------------------------------
 no.items.found = No items available
@@ -247,8 +265,6 @@ library.dialog.title =Library Manager
 
 library.dialog.hyperlink.release = Search repositories
 library.dialog.hyperlink.manually = Manually add Library from repository
-library.dialog.hyperlink.folder = Add root folder with *.class files
-library.dialog.hyperlink.tooltip = Use this option to specify the root folder under which to scan for controls.\nE.g. add the bin/ folder produced by your IDE during development of custom components.
 library.dialog.hyperlink.jar = Add Library/FXML from file system
 library.dialog.hyperlink.repositories = Manage repositories
